add utilities for working with asyncio and sudo
First we add functions to detect and prompt for a sudo password
OpenVPN runs with 'sudo -n', so it is useful to be able
to check whether a password would be necessary prior to this.

We also add a decorator that decorates async functions to turn
them into regular functions by running them in the default event loop.