improve hostname validation and rename 'normalize_hostname'
We prefer the noun 'normalized_hostname' because we are interested
in the return value of the function, not the way in which it alters
the state of the program.

The hostname validation checks as much as it can. This serves its
purpose as a preliminary check to avoid a roundtrip to the Nord API.